Using DI on 1 very small aspect - the webrootpath
I need the webrootpath to find my connection string.
It returns null in AppHarbor's environment for some reason.
I could get it from AppSettings, but then I'd have to hard-code the path
in a config file, which is kind of dumb.
Since there's no way to have a magical global variable without doing
something like service locator pattern, the interface type has to be
pushed down into CTORs from the controller on down to repositories. The
repositories are the only ones that actually need it ...